chore: remove `LegacySessionMetadata` and migrate web ui usage
