+++ multimodal/tarko/agent-web-ui/src/standalone/sidebar/ChatSession.tsx
@@ -87,10 +87,14 @@ export const ChatSession: React.FC<ChatSessionProps> = ({ isCollapsed }) => {
 
     return sessions.filter(
       (session) =>
-        session.name?.toLowerCase().includes(searchQuery.toLowerCase()) ||
-        session.tags?.some((tag) => tag.toLowerCase().includes(searchQuery.toLowerCase())),
+        session.metadata?.name?.toLowerCase().includes(searchQuery.toLowerCase()) ||
+        session.metadata?.tags?.some((tag) =>
+          tag.toLowerCase().includes(searchQuery.toLowerCase()),
+        ),
     );
   }, [sessions, searchQuery]);
+
+  console.log('filteredSessions', filteredSessions);
 
   // Optimize grouping calculation to reduce unnecessary re-computation
   const groupedSessions = useMemo(() => {
