+++ multimodal/tarko/interface/src/server.ts
@@ -115,3 +115,48 @@ export type AgioProviderConstructor<T extends AgentOptions = AgentOptions> = TCo
   AgioEvent.AgioProvider,
   [string, T, string, IAgent]
 >;
+
+/**
+ * Session metadata interface - JSON schema design for extensibility
+ */
+export interface SessionMetadata {
+  id: string;
+  createdAt: number;
+  updatedAt: number;
+  workspace: string;
+  // All extensible metadata in JSON format - no more schema migrations needed
+  metadata?: {
+    version?: number; // Schema version for backward compatibility
+    name?: string;
+    tags?: string[];
+    modelConfig?: {
+      provider: string;
+      modelId: string;
+      configuredAt: number;
+    };
+    agentConfig?: {
+      agentId: string;
+      configuredAt: number;
+      [key: string]: any; // Future agent configurations
+    };
+    [key: string]: any; // Future extensible fields
+  };
+}
+
+/**
+ * Legacy interface for backward compatibility during transition
+ * @deprecated Use SessionMetadata.metadata instead
+ */
+export interface LegacySessionMetadata {
+  id: string;
+  createdAt: number;
+  updatedAt: number;
+  name?: string;
+  workspace: string;
+  tags?: string[];
+  modelConfig?: {
+    provider: string;
+    modelId: string;
+    configuredAt: number;
+  };
+}
